# frontend/lib/api/openapi.json
{
 "components": {
  "schemas": {
...
   "ErrorBody": {
    "properties": {
     "details": {
      "title": "Details"
     },
     "error_code": {
      "title": "Error Code",
      "type": "string"
     },
     "message": {
      "title": "Message",
      "type": "string"
     },
     "request_id": {
      "title": "Request Id",
      "type": "string"
     },
     "trace_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Trace Id"
     }
    },
    "required": [
     "error_code",
     "message",
     "request_id"
    ],
    "title": "ErrorBody",
    "type": "object"
   },
...
   "ResolveRequest": {
    "additionalProperties": false,
    "description": "One customer conversation, oldest turn first, ending with the customer message to handle. Unknown fields (including any\nattempt to supply 'evidence', 'policy' or 'system' fields) are rejected: evidence only comes from the retrieval index.",
    "examples": [
     {
      "conversation": [
       {
        "role": "customer",
        "text": "My iPhone keeps changing \"it\" to \"I.T\" whenever I type. How do I fix this autocorrect bug?"
       }
      ],
      "metadata": {
       "channel": "twitter",
       "locale": "en-US"
      }
     }
    ],
    "properties": {
     "conversation": {
      "items": {
       "$ref": "#/components/schemas/TurnIn"
      },
      "maxItems": 200,
      "minItems": 1,
      "title": "Conversation",
      "type": "array"
     },
     "metadata": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/RequestMetadata"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "required": [
     "conversation"
    ],
    "title": "ResolveRequest",
    "type": "object"
   },
...
   "TraceListResponse": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/TraceSummary"
      },
      "title": "Items",
      "type": "array"
     },
     "limit": {
      "title": "Limit",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "limit"
    ],
    "title": "TraceListResponse",
    "type": "object"
   },
...
   "TurnIn": {
    "additionalProperties": false,
    "properties": {
     "role": {
      "description": "who wrote the turn; the LAST turn must be the customer's",
      "enum": [
       "customer",
       "brand"
      ],
      "title": "Role",
      "type": "string"
     },
     "text": {
      "description": "raw turn text; redacted by the agent before any storage or model call",
      "maxLength": 20000,
      "minLength": 1,
      "title": "Text",
      "type": "string"
     }
    },
    "required": [
     "role",
     "text"
    ],
    "title": "TurnIn",
    "type": "object"
   },
...
  }
 },
 "info": {
  "description": "Production-style reference implementation of an evidence-first AppleSupport agent.\n\n**One canonical endpoint:** `POST /api/v1/resolve` returns exactly one action (`AUTO_HANDLE`, `CLARIFICATION_REQUIRED`,\n`HUMAN_HANDOFF`) with the reason, the evidence, the verifier verdict, and the clarification or handoff packet.\nEvery execution is traced (`GET /api/v1/traces/{trace_id}`). Customer text is untrusted data: it is PII-redacted before any\nstorage or model call and can never act as an instruction or as evidence. When authentication is enabled, send\n`Authorization: Bearer <token>`; `/health` and `/ready` stay public.",
  "title": "ResolveAI API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
  "/api/v1/agent/profile": {
   "get": {
    "operationId": "agent_profile_api_v1_agent_profile_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Agent Profile Api V1 Agent Profile Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Not Found"
     },
     "411": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Length Required"
     },
     "413": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Content Too Large"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unprocessable Content"
     },
     "429": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Too Many Requests"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Internal Server Error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Service Unavailable"
     }
    },
    "summary": "Active agent configuration next to the evaluated configuration (read-only, no secrets)",
    "tags": [
     "operations"
    ]
   }
  },
  "/api/v1/config": {
   "get": {
    "operationId": "runtime_config_api_v1_config_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Runtime Config Api V1 Config Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Not Found"
     },
     "411": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Length Required"
     },
     "413": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Content Too Large"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unprocessable Content"
     },
     "429": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Too Many Requests"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Internal Server Error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Service Unavailable"
     }
    },
    "summary": "Safe runtime configuration (no secrets)",
    "tags": [
     "operations"
    ]
   }
  },
  "/api/v1/demo/scenarios": {
   "get": {
    "operationId": "demo_scenarios_api_v1_demo_scenarios_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "title": "Response Demo Scenarios Api V1 Demo Scenarios Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Not Found"
     },
     "411": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Length Required"
     },
     "413": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Content Too Large"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unprocessable Content"
     },
     "429": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Too Many Requests"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Internal Server Error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Service Unavailable"
     }
    },
    "summary": "Synthetic demo scenarios and their expected outcomes",
    "tags": [
     "evaluation"
    ]
   }
  },
  "/api/v1/evaluation/release": {
   "get": {
    "operationId": "evaluation_release_api_v1_evaluation_release_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Evaluation Release Api V1 Evaluation Release Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Not Found"
     },
     "411": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Length Required"
     },
     "413": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Content Too Large"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unprocessable Content"
     },
     "429": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Too Many Requests"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Internal Server Error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Service Unavailable"
     }
    },
    "summary": "Release 1.0.0 scorecard from the frozen release artifacts, by dataset",
    "tags": [
     "evaluation"
    ]
   }
  },
  "/api/v1/evaluation/summary": {
   "get": {
    "operationId": "evaluation_summary_api_v1_evaluation_summary_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Evaluation Summary Api V1 Evaluation Summary Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Not Found"
     },
     "411": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Length Required"
     },
     "413": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Content Too Large"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unprocessable Content"
     },
     "429": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Too Many Requests"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Internal Server Error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Service Unavailable"
     }
    },
    "summary": "Headline evaluation results from the frozen Phase 6 artifacts",
    "tags": [
     "evaluation"
    ]
   }
  },
...
  "/api/v1/knowledge/summary": {
   "get": {
    "operationId": "knowledge_summary_api_v1_knowledge_summary_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Knowledge Summary Api V1 Knowledge Summary Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Not Found"
     },
     "411": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Length Required"
     },
     "413": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Content Too Large"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unprocessable Content"
     },
     "429": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Too Many Requests"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Internal Server Error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Service Unavailable"
     }
    },
    "summary": "Knowledge-base aggregates (counts and shares only, no message text)",
    "tags": [
     "operations"
    ]
   }
  },
...
  "/api/v1/resolve": {
   "post": {
    "description": "Runs the full pipeline (PII redaction, injection check, intent, retrieval, evidence gate, risk flags, deterministic policy, draft / clarify / handoff, verification, output gate) and returns what happened, why, on which evidence, and what happens next. Exactly one action is returned: AUTO_HANDLE, CLARIFICATION_REQUIRED or HUMAN_HANDOFF. A model outage, a model timeout or an exhausted request time budget is not an HTTP error: the agent's deterministic fallback returns a safe clarification or handoff.",
    "operationId": "resolve_api_v1_resolve_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ResolveRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ResolveResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Not Found"
     },
     "411": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Length Required"
     },
     "413": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Content Too Large"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unprocessable Content"
     },
     "429": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Too Many Requests"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Internal Server Error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Service Unavailable"
     }
    },
    "summary": "Run one conversation through the ResolveAI agent",
    "tags": [
     "agent"
    ]
   }
  },
  "/api/v1/traces": {
   "get": {
    "operationId": "list_traces_api_v1_traces_get",
    "parameters": [
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 20,
       "maximum": 200,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "description": "any capitalization",
      "in": "query",
      "name": "action",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "pattern": "^(?i)(AUTO_HANDLE|CLARIFICATION_REQUIRED|HUMAN_HANDOFF)$",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "any capitalization",
       "title": "Action"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TraceListResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Not Found"
     },
     "411": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Length Required"
     },
     "413": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Content Too Large"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Unprocessable Content"
     },
     "429": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Too Many Requests"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Internal Server Error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody"
        }
       }
      },
      "description": "Service Unavailable"
     }
    },
    "summary": "Most recent traces (newest first)",
    "tags": [
     "audit"
    ]
   }
  },
...
 },
...
}